FrontEnd/Backend/requirements.txt
aiohttp>=3.8.0
websockets>=10.0
colorama>=0.4.4
solders>=0.9.0
python-dotenv>=0.19.0
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.2
httpx>=0.23.0
python-multipart>=0.0.5
typing-extensions>=4.0.0
asyncio>=3.4.3